Add effort spreader
These changes adds the effort spreader function. This function takes a maximum effort emax, a switch for capping the effort at emax, and a vector of open months (like the season vector previously). This is passed to eff_spread which allocates the effort appropriately and, with e_cap = FALSE, ramps effort up/down depending on the # of months open. Also added general comments.